feat: refactor database structure and enhance API functionality
- Introduced a new database module to centralize database interactions.
- Updated import paths across various controllers to reference the new database module.
- Added new dependencies for cryptographic functions and encoding.
- Enhanced TypeScript configuration for better module resolution and source mapping.
- Updated package versions in pnpm-lock.yaml for improved compatibility.
