+++ apps/hono-api/src/activity/controllers/create-activity.ts
@@ -1,5 +1,5 @@
-import db from "../../../database";
-import { activityTable } from "../../../database/schema";
+import db from "../../database";
+import { activityTable } from "../../database/schema";
 
 async function createActivity(
   taskId: string,
