+++ apps/hono-api/src/index.ts
@@ -3,8 +3,8 @@ import { migrate } from "drizzle-orm/better-sqlite3/migrator";
 import { Hono } from "hono";
 import { getCookie } from "hono/cookie";
 import { cors } from "hono/cors";
-import db from "../database";
 import activity from "./activity";
+import db from "./database";
 import { auth } from "./middlewares/auth";
 import project from "./project";
 import task from "./task";
