+++ apps/hono-api/src/project/controllers/get-project.ts
@@ -1,6 +1,6 @@
 import { and, eq } from "drizzle-orm";
-import db from "../../../database";
-import { projectTable, workspaceTable } from "../../../database/schema";
+import db from "../../database";
+import { projectTable, workspaceTable } from "../../database/schema";
 
 async function getProject(id: string, workspaceId: string) {
   const [project] = await db
