+++ apps/hono-api/src/task/controllers/get-task.ts
@@ -1,7 +1,7 @@
 import { eq } from "drizzle-orm";
 import { HTTPException } from "hono/http-exception";
-import db from "../../../database";
-import { taskTable } from "../../../database/schema";
+import db from "../../database";
+import { taskTable } from "../../database/schema";
 
 async function getTask(taskId: string) {
   const task = await db.query.taskTable.findFirst({
