+++ apps/hono-api/src/task/controllers/update-task-status.ts
@@ -1,6 +1,6 @@
 import { eq } from "drizzle-orm";
-import db from "../../../database";
-import { taskTable } from "../../../database/schema";
+import db from "../../database";
+import { taskTable } from "../../database/schema";
 
 async function updateTaskStatus({
   id,
