+++ apps/hono-api/src/task/controllers/update-task.ts
@@ -1,7 +1,7 @@
 import { eq } from "drizzle-orm";
 import { HTTPException } from "hono/http-exception";
-import db from "../../../database";
-import { taskTable } from "../../../database/schema";
+import db from "../../database";
+import { taskTable } from "../../database/schema";
 
 async function updateTask(
   taskId: string,
