+++ apps/hono-api/src/user/controllers/sign-up.ts
@@ -1,7 +1,7 @@
 import bcrypt from "bcrypt";
 import { HTTPException } from "hono/http-exception";
-import db from "../../../database";
-import { userTable } from "../../../database/schema";
+import db from "../../database";
+import { userTable } from "../../database/schema";
 import { publishEvent } from "../../events";
 import getSettings from "../../utils/get-settings";
