+++ apps/hono-api/src/user/utils/create-session.ts
@@ -1,7 +1,7 @@
 import { sha256 } from "@oslojs/crypto/sha2";
 import { encodeHexLowerCase } from "@oslojs/encoding";
-import db from "../../../database";
-import { sessionTable } from "../../../database/schema";
+import db from "../../database";
+import { sessionTable } from "../../database/schema";
 
 async function createSession(token: string, userId: string) {
   const sessionId = encodeHexLowerCase(sha256(new TextEncoder().encode(token)));
