+++ apps/hono-api/src/user/utils/invalidate-session.ts
@@ -1,6 +1,6 @@
 import { eq } from "drizzle-orm";
-import db from "../../../database";
-import { sessionTable } from "../../../database/schema";
+import db from "../../database";
+import { sessionTable } from "../../database/schema";
 
 async function invalidateSession(sessionId: string) {
   await db.delete(sessionTable).where(eq(sessionTable.id, sessionId));
