+++ apps/hono-api/src/utils/purge-demo-data.ts
@@ -1,10 +1,10 @@
-import { projectTable } from "../../database/schema";
-import { taskTable } from "../../database/schema";
-import { userTable } from "../../database/schema";
-import { workspaceTable } from "../../database/schema";
-import { workspaceUserTable } from "../../database/schema";
+import { projectTable } from "../database/schema";
+import { taskTable } from "../database/schema";
+import { userTable } from "../database/schema";
+import { workspaceTable } from "../database/schema";
+import { workspaceUserTable } from "../database/schema";
 
-import db from "../../database";
+import db from "../database";
 
 async function purgeData() {
   await db.delete(userTable);
