+++ apps/hono-api/src/workspace-user/controllers/get-workspace-users.ts
@@ -1,6 +1,6 @@
 import { asc, eq } from "drizzle-orm";
-import db from "../../../database";
-import { userTable, workspaceUserTable } from "../../../database/schema";
+import db from "../../database";
+import { userTable, workspaceUserTable } from "../../database/schema";
 
 function getWorkspaceUsers(workspaceId: string) {
   return db
