+++ apps/hono-api/src/workspace/controllers/delete-workspace.ts
@@ -1,7 +1,7 @@
 import { and, eq } from "drizzle-orm";
 import { HTTPException } from "hono/http-exception";
-import db from "../../../database";
-import { workspaceTable } from "../../../database/schema";
+import db from "../../database";
+import { workspaceTable } from "../../database/schema";
 
 async function deleteWorkspace(userEmail: string, workspaceId: string) {
   const [existingWorkspace] = await db
